+++ app/src/main/java/de/tap/easy_xkcd/mainActivity/MainActivityViewModel.kt
@@ -21,10 +21,6 @@ class MainActivityViewModel @Inject constructor(
 ) : AndroidViewModel(app) {
 
     fun onCreateWithNullSavedInstanceState() {
-        viewModelScope.launch {
-            repository.migrateRealmDatabase()
-        }
-
         newComicNotificationHandler.initNotifications()
     }
 }